+++ packages/fx-core/tests/component/generator/officeAddinGenerator.test.ts
@@ -27,7 +27,7 @@ import * as path from "path";
 import proxyquire from "proxyquire";
 import * as sinon from "sinon";
 import * as uuid from "uuid";
-import { cpUtils } from "../../../src/common/deps-checker";
+import { cpUtils } from "../../../src/component/deps-checker/";
 import { manifestUtils } from "../../../src/component/driver/teamsApp/utils/ManifestUtils";
 import { Generator } from "../../../src/component/generator/generator";
 import {
