+++ packages/fx-core/tests/component/generator/spfxGenerator.test.ts
@@ -18,7 +18,7 @@ import mockedEnv, { RestoreFn } from "mocked-env";
 import * as path from "path";
 import * as sinon from "sinon";
 import * as uuid from "uuid";
-import { cpUtils } from "../../../src/common/deps-checker";
+import { cpUtils } from "../../../src/component/deps-checker/";
 import { ManifestUtils } from "../../../src/component/driver/teamsApp/utils/ManifestUtils";
 import { Generator } from "../../../src/component/generator/generator";
 import { GeneratorChecker } from "../../../src/component/generator/spfx/depsChecker/generatorChecker";
